Refactor test to work
