--- mysocial/authors/models/__init__.py
@@ -1,1 +1,0 @@
-from .author import Author
