--- mysocial/common/__init__.py
@@ -1,2 +1,0 @@
-from .pagination_helper import PaginationHelper
-from .test_helper import TestHelper
