+++ mysocial/common/test_helper.py
@@ -1,4 +1,4 @@
-from authors.models import Author
+from authors.models.author import Author
 
 
 class TestHelper:
